pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = ""
authors = ["Joshua Goss"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.66.0"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
SQLAlchemy = {extras = ["postgresql"], version = "^1.4.21"}
python-dotenv = "^0.18.0"
pydantic = "^1.8.2"
alembic = "^1.6.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^1.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
